requirements.txt
langgraph>=0.0.15
langchain>=0.1.0
langchain-openai>=0.0.2
pydantic>=2.0.0
python-dotenv>=1.0.0
C:/PythonProjects/genAI_utils -e